hide moves to previous collectible
